Add row-setter regression tests, split into UpdateFrom.Row.cs
