+++ Tests/Base/Asserts/ShouldlyMissingExtensions.cs
@@ -2,8 +2,6 @@
 using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 
-using NUnit.Framework;
-
 namespace Shouldly
 {
 	public static class ShouldlyMissingExtensions
@@ -11,17 +9,13 @@ namespace Shouldly
 		[MethodImpl(MethodImplOptions.NoInlining)]
 		public static void ShouldContain(this string actual, string expected, ITimesConstraint times)
 		{
-			if (times.Type == TimesType.Exactly)
+			var count = CountInString(actual, expected);
+
+			switch (times.Type)
 			{
-				Assert.That(CountInString(actual, expected), Is.EqualTo(times.Times));
-			}
-			else if (times.Type == TimesType.AtLeast)
-			{
-				Assert.That(CountInString(actual, expected), Is.GreaterThanOrEqualTo(times.Times));
-			}
-			else
-			{
-				Assert.Fail($"Unknown times constraint: {times.Type}");
+				case TimesType.Exactly: count.ShouldBe(times.Times);                   break;
+				case TimesType.AtLeast: count.ShouldBeGreaterThanOrEqualTo(times.Times); break;
+				default: throw new ShouldAssertException($"Unknown times constraint: {times.Type}");
 			}
 
 			static int CountInString(string str, string fragment)
@@ -41,7 +35,7 @@ namespace Shouldly
 		public static void ShouldNotContainAny(this string actual, params string[] notFound)
 		{
 			foreach (var str in notFound)
-				Assert.That(actual, Does.Not.Contain(str));
+				actual.ShouldNotContain(str);
 		}
 
 		public static void ShouldAllSatisfy<T>(this IEnumerable<T> collection, Action<T> asserts)
