+++ Tests/Linq/Update/UpdateFromTests.cs
@@ -13,7 +13,7 @@ using Shouldly;
 namespace Tests.xUpdate
 {
 	[TestFixture]
-	public class UpdateFromTests : TestBase
+	public partial class UpdateFromTests : TestBase
 	{
 		[Table]
 		public partial class UpdatedEntities
@@ -605,125 +605,6 @@ namespace Tests.xUpdate
 						Value3 = x.Value2,
 						Value4 = x.Value2,
 					});
-		}
-
-		[Table]
-		sealed class UpdateSubquerySourceTable
-		{
-			[PrimaryKey] public int Id { get; set; }
-			[Column] public string? FirstName { get; set; }
-			[Column] public string? LastName { get; set; }
-		}
-
-		[Test(Description = "https://github.com/linq2db/linq2db/issues/5413")]
-		public void UpdateFromSubqueryRowShouldBeOptimized([IncludeDataSources(TestProvName.AllSQLite, TestProvName.AllOracle, TestProvName.AllPostgreSQL, TestProvName.AllInformix)] string context)
-		{
-			using var db = GetDataContext(context);
-
-			using var _ = new DeletePerson(db);
-
-			using var sourceTable = db.CreateLocalTable(
-			[
-				new UpdateSubquerySourceTable { Id = 1, FirstName = "FirstTooth", LastName = "FirstFairy" },
-				new UpdateSubquerySourceTable { Id = 2, FirstName = "SecondTooth", LastName = "SecondFairy" },
-				new UpdateSubquerySourceTable { Id = 3, FirstName = "ThirdTooth", LastName = "ThirdFairy" }
-			]);
-
-			var affectedCount = sourceTable
-				.Where(x => x.Id == 1)
-				.Set(
-					x => Sql.Row(x.FirstName, x.LastName),
-					x => (
-						from s in db.SelectQuery(() => 1)
-						from canChange in sourceTable.Where(t => t.Id == x.Id + 1).DefaultIfEmpty()
-						select Sql.Row(
-							canChange != null ? canChange.FirstName! : x.FirstName,
-							canChange != null ? canChange.LastName!  : x.LastName
-						)
-					).Single()
-				)
-				.Update();
-
-			affectedCount.ShouldBe(1);
-
-			var records = sourceTable.OrderBy(x => x.Id).ToList();
-
-			records[0].FirstName.ShouldBe("SecondTooth");
-			records[0].LastName.ShouldBe("SecondFairy");
-
-			records[1].FirstName.ShouldBe("SecondTooth");
-			records[1].LastName.ShouldBe("SecondFairy");
-
-			records[2].FirstName.ShouldBe("ThirdTooth");
-			records[2].LastName.ShouldBe("ThirdFairy");
-		}
-
-		[Test(Description = "Row-setter with mixed literal + correlated values (regression for ProcessUpdateItemsWithRows column/value pairing)")]
-		public void UpdateFromSubqueryRowMixedIndependentAndDependent([IncludeDataSources(TestProvName.AllSQLite, TestProvName.AllOracle, TestProvName.AllPostgreSQL)] string context)
-		{
-			using var db = GetDataContext(context);
-
-			using var sourceTable = db.CreateLocalTable(
-			[
-				new UpdateSubquerySourceTable { Id = 1, FirstName = "FirstTooth",  LastName  = "FirstFairy"  },
-				new UpdateSubquerySourceTable { Id = 2, FirstName = "SecondTooth", LastName  = "SecondFairy" },
-				new UpdateSubquerySourceTable { Id = 3, FirstName = "ThirdTooth",  LastName  = "ThirdFairy"  }
-			]);
-
-			// (FirstName, LastName) = ("literal", (subquery on next row).LastName)
-			// FirstName side is independent of other tables; LastName side depends on a
-			// correlated subquery. Exercises the independent/dependent split in
-			// ProcessUpdateItemsWithRows — earlier code emitted (col, col) = (col, col)
-			// for the independent slot, dropping the literal value.
-			var affectedCount = sourceTable
-				.Where(x => x.Id == 1)
-				.Set(
-					x => Sql.Row(x.FirstName, x.LastName),
-					x => Sql.Row(
-						(string?)"literalFirst",
-						sourceTable.Where(t => t.Id == x.Id + 1).Select(t => t.LastName).First()))
-				.Update();
-
-			affectedCount.ShouldBe(1);
-
-			var records = sourceTable.OrderBy(x => x.Id).ToList();
-
-			records[0].FirstName.ShouldBe("literalFirst");
-			records[0].LastName .ShouldBe("SecondFairy");
-
-			records[1].FirstName.ShouldBe("SecondTooth");
-			records[1].LastName .ShouldBe("SecondFairy");
-
-			records[2].FirstName.ShouldBe("ThirdTooth");
-			records[2].LastName .ShouldBe("ThirdFairy");
-		}
-
-		[Test(Description = "https://github.com/linq2db/linq2db/issues/5413")]
-		public void UpdateFromSubqueryRowShouldRemainSimple([IncludeDataSources(TestProvName.AllOracle)] string context)
-		{
-			using var db = GetDataContext(context);
-			using var table1 = db.CreateLocalTable<NewEntities>();
-			using var table2 = db.CreateLocalTable<UpdatedEntities>();
-			using var table3 = db.CreateLocalTable<UpdateRelation>();
-
-			table1
-				.Where(u1 => u1.id == 7)
-				.Set(
-					u1 => Sql.Row(u1.Value1, u1.Value2),
-					u1 => (
-						from c in db.SelectQuery(() => u1.Value3 + 10)
-						from n2 in table2.LeftJoin(n2 => n2.id == c)
-						from n3 in table3.LeftJoin(n3 => n2.RelationId == n3.id)
-						where n3.RelatedValue3 < 1000
-						select Sql.Row(n2.Value1, n3.RelatedValue2))
-						.Single()
-				)
-				.Update();
-
-			LastQuery!.ShouldContain("\"NewEntities\"",     Exactly.Once());
-			LastQuery!.ShouldContain("\"UpdatedEntities\"", Exactly.Once());
-			LastQuery!.ShouldContain("\"UpdateRelation\"",  Exactly.Once());
-			LastQuery!.ShouldNotContain("EXISTS");
 		}
 
 		[ThrowsForProvider(typeof(LinqToDBException), TestProvName.AllClickHouse, ErrorMessage = ErrorHelper.ClickHouse.Error_CorrelatedUpdate)]
